requirements_file.txt
patchright>=0.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0

# AI Providers (optional)
g4f>=0.2.0
openai>=1.0.0

# GUI
Pillow>=10.0.0